+++ tomcat/src/main/java/com/techcourse/controller/LoginController.java
@@ -6,11 +6,11 @@ import java.io.IOException;
 import java.util.Optional;
 import org.apache.catalina.Manager;
 import org.apache.catalina.controller.AbstractController;
-import org.apache.coyote.http11.HttpCookie;
-import org.apache.coyote.http11.HttpRequest;
-import org.apache.coyote.http11.HttpResponse;
-import org.apache.coyote.http11.Session;
-import org.apache.coyote.http11.SessionManager;
+import org.apache.coyote.http11.request.HttpCookie;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
+import org.apache.coyote.http11.session.Session;
+import org.apache.coyote.http11.session.SessionManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
