+++ tomcat/src/main/java/com/techcourse/controller/StaticResourceController.java
@@ -3,10 +3,10 @@ package com.techcourse.controller;
 import java.io.IOException;
 import java.io.InputStream;
 import org.apache.catalina.controller.AbstractController;
-import org.apache.coyote.http11.ContentType;
-import org.apache.coyote.http11.HttpRequest;
-import org.apache.coyote.http11.HttpResponse;
-import org.apache.coyote.http11.HttpStatus;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.ContentType;
+import org.apache.coyote.http11.response.HttpResponse;
+import org.apache.coyote.http11.response.HttpStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
