+++ tomcat/src/main/java/org/apache/catalina/controller/Controller.java
@@ -1,7 +1,7 @@
 package org.apache.catalina.controller;
 
-import org.apache.coyote.http11.HttpRequest;
-import org.apache.coyote.http11.HttpResponse;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
 
 public interface Controller {
     void service(HttpRequest request, HttpResponse response) throws Exception;
