+++ tomcat/src/main/java/org/apache/catalina/controller/FrontController.java
@@ -4,8 +4,8 @@ import com.techcourse.controller.HelloController;
 import com.techcourse.controller.LoginController;
 import com.techcourse.controller.RegisterController;
 import com.techcourse.controller.StaticResourceController;
-import org.apache.coyote.http11.HttpRequest;
-import org.apache.coyote.http11.HttpResponse;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
 
 public class FrontController {
