+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -5,6 +5,9 @@ import java.io.IOException;
 import java.net.Socket;
 import org.apache.catalina.controller.FrontController;
 import org.apache.coyote.Processor;
+import org.apache.coyote.http11.request.HttpRequest;
+import org.apache.coyote.http11.response.HttpResponse;
+import org.apache.coyote.http11.response.HttpStatus;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
